refactor move folders from inside ml_potentials to outside ml_potentials
